requirements.txt
streamlit>=1.28
streamlit-mermaid
